live_trading_system/requirements.txt
fastapi
uvicorn
websockets
python-binance
pandas